chore: update

+++ spacevim.mk
@@ -6,7 +6,6 @@ include $(dotmk)/dotmk.mk
 include $(dotmk)/brew.mk
 include $(dotmk)/node.mk
 include $(dotmk)/conflate.mk
-include $(dotmk)/yarn.mk
 
 .PHONY: \
 	install \
@@ -15,19 +14,13 @@ include $(dotmk)/yarn.mk
 install: install.spacevim
 install.spacevim: \
 	install.conflate \
-	install.yarn \
 	$(BREW_HOME)/Cellar/neovim \
 	$(BREW_HOME)/Cellar/msgpack \
 	$(BREW_HOME)/Cellar/python@3.9 \
 	$(BREW_HOME)/Cellar/global \
 	$(BREW_HOME)/Cellar/moreutils \
 	$(HOME)/.SpaceVim \
-	.SpaceVim.d/init.toml \
-# 	$(yarn.global.mod)/import-js \
-# 	$(yarn.global.mod)/javascript-typescript-langserver \
-# 	$(yarn.global.mod)/typescript \
-# 	$(yarn.global.mod)/javascript-typescript-langserver \
-# 	$(yarn.global.mod)/typescript-language-server 
+	.SpaceVim.d/init.toml
 install.spacevim:
 install.spacevim:
 	(which bash-language-server | grep -v 'not found') || volta install bash-language-server
@@ -39,32 +32,44 @@ install.spacevim:
 
 trash: trash.spacevim
 trash.spacevim:
+ifeq ($(CURDIR),$(HOME))
 	curl -sLf https://spacevim.org/install.sh | bash -s -- --uninstall
 	rm -rf ~/.SpaceVim
 	rm -rf ~/.cache/SpaceVim
 	rm -rf ~/.cache/vimfiles
+endif
+	rm -rf .SpaceVim.d
 
 $(HOME)/.SpaceVim:
 	curl -sLf https://spacevim.org/install.sh | bash
 
-$(HOME)/.SpaceVim.d/custom.toml:
-	mkdir -p $(dir $@)
-	touch $@
+$(HOME)/.SpaceVim.d/init.toml:
+	[ -e $@ ] || (mkdir $(dir $@) && touch $@)
+
+$(CURDIR)/.SpaceVim.d/init.toml:
+	[ -e $@ ] || (mkdir $(dir $@) && touch $@)
 
 .SpaceVim.d/init.toml: 
-	$(eval %/.SpaceVim.d/.custom.yml := \
+	$(eval %/.SpaceVim.d/.init.yml := \
 		$(foreach pre,$^,$(patsubst $(dir $(pre))%.toml,$(dir $(pre)).%.yml,$(pre))))
-	make $(%/.SpaceVim.d/.custom.yml) 
+	make $(%/.SpaceVim.d/.init.yml)
 	conflate \
-		$(patsubst %,-data %,$(%/.SpaceVim.d/.custom.yml)) \
-		-data .SpaceVim.d/.custom.yml \
+		$(patsubst %,-data %,$(%/.SpaceVim.d/.init.yml)) \
 		-format TOML \
 		| sponge $@
 
-%/.SpaceVim.d/.custom.yml: %/.SpaceVim.d/custom.toml
-	conflate -data $^ -format YAML > $@ 
+%/.SpaceVim.d/.init.yml: %/.SpaceVim.d/init.toml
+	conflate -expand -data $^ -format YAML > $@
 
-.SpaceVim.d/init.toml: $(dotmk)/spacevim/.SpaceVim.d/custom.toml 
-.SpaceVim.d/init.toml: $(HOME)/.SpaceVim.d/custom.toml
+# Default DotMK config
+.SpaceVim.d/init.toml: $(dotmk)/spacevim/.SpaceVim.d/init.toml 
+# Add custom global config for user in projects
+ifneq ($(CURDIR),$(HOME))
+.SpaceVim.d/init.toml: $(HOME)/.SpaceVim.d/init.toml
+endif
+# Override default config if init.toml was already present
+.SpaceVim.d/init.toml: $(CURDIR)/.SpaceVim.d/init.toml
+
+.gitignore: $(dotmk)/spacevim/.gitignore
 
 endif # spacevim.mk
